+++ src/main/java/com/logistics/mlogistics/service/SupplyOrderItemService.java
@@ -42,7 +42,9 @@ public class SupplyOrderItemService {
         });
     }
 
-    public void delete(UUID id) {
+    public boolean delete(UUID id) {
+        if (!repository.existsById(id)) return false;
         repository.deleteById(id);
+        return true;
     }
 }
